remake `/products` page
