+++ src/app/products/page.tsx
@@ -3,8 +3,72 @@ export const metadata: Metadata = {
   title: 'Продукты',
 }
 
+import {client} from '@/lib/sanity'
+import {revalidateOnTime} from '@/lib/utils'
+
+import Container from '#/UI/Container'
 import Catalog from '##/products/Catalog'
 
-export default function Products() {
-  return <Catalog />
+export interface Product {
+  name: string
+  caption: string
+  article: number
+  volume: string
+  description: string
+  usage: string
+  price: number
+  slug: {current: string}
+  image?: Array<{asset: {url: string}}>
+  main_filter: string
+  for_face?: string
+  for_body?: string
+  product_type: string
+  by_intention: string[]
+  product_line: string
+  [x: string]: any
 }
+
+async function getData(): Promise<Product[]> {
+  const data = await client.fetch<Product>(
+    `*[_type == 'product' && slug.current != null] {
+        name,
+        caption,
+        article,
+        volume,
+        description,
+        usage,
+        price,
+        slug,
+        image,
+        main_filter,
+        for_face,
+        for_body,
+        product_type,
+        by_intention,
+        product_line,
+    }`,
+    {},
+    {
+      next: {
+        revalidate: revalidateOnTime,
+      },
+    },
+  )
+  return Array.isArray(data) ? data : []
+}
+
+const ProductsPage = async () => {
+  const products: Product[] = await getData()
+
+  if (!products) {
+    return <mark>Произошла ошибка при получении данных!</mark>
+  }
+
+  return (
+    <Container className="w-[80%]">
+      <Catalog products={products} />
+    </Container>
+  )
+}
+
+export default ProductsPage
